pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ariadne-backend"
version = "0.1.0"
description = "Ariadne metadata management backend (FastAPI)"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi==0.115.3",
    "uvicorn[standard]==0.32.0",
    "python-multipart>=0.0.19",
    "sqlalchemy[asyncio]==2.0.36",
    "alembic==1.15.0",
    "asyncpg==0.31.0",
    "pgvector==0.3.2",
    "neo4j==5.28.2",
    "redis==5.2.1",
    "pydantic==2.10.0",
    "pydantic-settings==2.7.0",
    "email-validator==2.2.0",
    "pyjwt>=2.10.0",
    "passlib[bcrypt]==1.7.4",
    "structlog==25.1.0",
    "openpyxl==3.1.5",
    "pyyaml==6.0.2",
    "python-dotenv==1.0.1",
    "bcrypt==4.2.0",
    "schemathesis==4.7.5",
    "ruff==0.14.2",
    "black==23.11.0",
    "mypy==1.13.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "httpx==0.28.1",
    "types-redis>=4.6.0.20241004",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.setuptools]
package-dir = {"" = "backend"}
packages = ["app"]

[project.optional-dependencies]
dev = [
    "ruff==0.14.2",
    "black==23.11.0",
    "mypy==1.13.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "httpx==0.28.1",
    "types-redis",
    "types-PyYAML",
    "schemathesis==4.7.5",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
